fix: carry fractional scroll remainder across flushes
ScrollSession truncated each computed scroll amount to int before handing it
to the scroller, while advancing the contact position that earned it, so the
sub-unit fraction was discarded every event. The loss was systematic and
scaled with 1/vSens: ~5% at vSens=15, ~29% at low sensitivity, making
effective sensitivity quietly lower than the configured value.

Scroller::Scroll now takes a double and WinScroller accumulates fractionally,
emitting whole units and retaining the remainder for the next flush. This
also makes vSens (gain) and scrollFlushMs (emit rate) properly orthogonal.

+++ ChiralScroll/src/Scroller.h
@@ -9,7 +9,10 @@ public:
 	virtual ~Scroller() = default;
 
 	virtual void StartScrolling() = 0;
-	virtual void Scroll(int amt) = 0;
+	// Scrolls by amt wheel units. Fractional amounts are the caller's real
+	// computed value; implementations accumulate them rather than truncating,
+	// so slow movement is not systematically lost.
+	virtual void Scroll(double amt) = 0;
 	virtual void StopScrolling() = 0;
 };
